plumbed the input/output into MINA and filled out more of the encoding/decoding
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/qpid/trunk@561665 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/common/src/main/java/org/apache/qpidity/TrackSwitch.java
@@ -28,10 +28,10 @@ package org.apache.qpidity;
  * @author Rafael H. Schloming
  */
 
-class TrackSwitch<C> extends Switch<Short,Event<C,Frame>>
+class TrackSwitch<C> extends Switch<Byte,Event<C,Frame>>
 {
 
-    public Short resolve(Event<C,Frame> event)
+    public Byte resolve(Event<C,Frame> event)
     {
         return event.target.getTrack();
     }
